계좌출금 서비스

+++ src/main/java/com/bank/domain/account/Account.java
@@ -65,7 +65,7 @@ public class Account {
     }
 
     public void checkSamePassword(Long password) {
-        if (this.password != password) {
+        if (this.password.longValue() != password.longValue()) {
             throw new CustomApiException("계좌 비밀번호 검증에 실패했습니다.");
         }
     }
